[TESTING] more stablization of dev-v6.0 (- WIP #316 -)
* renamed a few modules to make linters calm down more
* slight improvement for imports
* reconfigured dependabot slightly to distinguish "lock-files" (those used as such anyway) from declared project dependencies
