+++ piplicenses/output/_prettytable_bridge.py
@@ -1,0 +1,44 @@
+# vim:fenc=utf-8 ff=unix ft=python ts=4 sw=4 sts=4 si et
+"""
+pip-licenses.output._prettytable_bridge
+
+MIT-0 License OR MIT License
+
+Rationale: if the license is bigger than the actual code, then MIT-0 is
+sufficient, and compatible to just the MIT-0, so we'll be explicit here.
+
+---
+
+MIT No Attribution (MIT-0 License)
+
+Copyright (c) 2026 Mr. Walls
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+import prettytable
+from prettytable import (
+    HRuleStyle,
+    PrettyTable,
+    RowType,
+)
+
+__all__ = [
+    """HRuleStyle""",
+    """PrettyTable""",  # the class
+    """RowType""",
+    """prettytable""",  # the module
+]
