+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=82.0.0", "setuptools_scm"]
+requires = ["setuptools>=82.0.1", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -10,7 +10,7 @@ requires-python = ">=3.9"
 license = "MIT"
 license-files = ["LICENSE"]
 authors = [
-	{"name" = "raimon", "email" = "raimon49@hotmail.com"}, # 2018-2025 (retired)
+	{"name" = "raimon", "email" = "raimon49@users.noreply.github.com"}, # 2018-2025 (retired)
 	{"name" = "Mr. Walls", "email" = "reactive-firewall@users.noreply.github.com"},
 ]
 maintainers = [
@@ -62,15 +62,20 @@ output-file = "dev-requirements.in"
 [project.urls]
 homepage = "https://github.com/raimon49/pip-licenses"
 issues = "https://github.com/raimon49/pip-licenses/issues"
+license = "https://raw.githubusercontent.com/raimon49/pip-licenses/HEAD/LICENSE"
 releasenotes = "https://github.com/raimon49/pip-licenses/releases"
 repository = "https://github.com/raimon49/pip-licenses.git"
+
+#[project.readme]
+#file = "README.md"
+#content-type = "text/markdown; charset=UTF-8; variant=GFM"
 
 [tool.setuptools]
 py-modules = ["pip-licenses"]
 
 [tool.setuptools.dynamic]
 version = {attr = "piplicenses.__version__"}
-readme = {file = ["README.md", "CHANGELOG.md"], content-type = "text/markdown"}
+readme = {file = ["README.md", "CHANGELOG.md"], content-type = "text/markdown; charset=UTF-8; variant=GFM"}
 
 [tool.pytest.ini_options]
 addopts = "-v --cov --cov-report term-missing"
@@ -105,13 +110,12 @@ extend-select = [
 extend-ignore = [
 	"C401",   # Unnecessary generator (rewrite as a set comprehension)
 	"F841",   # Local variable is assigned to but never used
+	"FA100",  # DEPRECATED in v6.0+  # See new typing PEPs (TODO: enumerate)
 	"SIM300", # Yoda condition detected
 	"RUF005", # Consider unpacking instead of concatenation
 	"RUF012", # Unnecessary noise about Python language (In Python defaults are always mutable)  # See https://github.com/astral-sh/ruff/issues/5243
 	"RUF021", # Parenthesize `a and b` expressions when chaining `and` and `or` together, to make the precedence clear
 	"RUF013", # Ignore PEP484 until GHI-316 is resolved
-	"FA100",  # DEPRECATED in v6.0+  # See new
-	"N999",   # Ignore until GHI-316 is resolved
 ]
 isort.known-first-party = ["piplicenses"]
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?)(_{2,2})?)$"  # See https://regex101.com/r/U0tpDT/2
